Make ContigId u32 instead of u16

+++ src/command/prune.rs
@@ -238,8 +238,8 @@ fn load_divergences(
     }
     if let Some(k) = missing_k {
         let (i, j) = divergences.from_linear_index(k);
-        let name1 = contigs.get_name(ContigId::new(i as u16));
-        let name2 = contigs.get_name(ContigId::new(j as u16));
+        let name1 = contigs.get_name(ContigId::new(i));
+        let name2 = contigs.get_name(ContigId::new(j));
         log::warn!("[{}] Divergence missing for {:.1}% haplotypes pairs, for example {} and {}. Replacing with {:.5}",
             contigs.tag(), 100.0 * missing_count as f64 / divergences.linear_len() as f64, name1, name2, repl_missing);
     }
